+++ docs/book/src/simple-external-plugin-tutorial/flags.go
@@ -1,0 +1,70 @@
+package main
+
+import (
+	"github.com/spf13/pflag"
+	"sigs.k8s.io/kubebuilder/v3/pkg/plugin/external"
+)
+
+// flagsCmd handles all the logic for the `flags` subcommand of the sample external plugin.
+// In Kubebuilder's Phase 2 Plugins the `flags` subcommand is an optional subcommand for
+// external plugins to support. The `flags` subcommand allows for an external plugin
+// to provide Kubebuilder with a list of flags that the `init`, `create api`, `create webhook`,
+// and `edit` subcommands allow. This allows Kubebuilder to give an external plugin the ability
+// to feel like a native Kubebuilder plugin to a Kubebuilder user by only binding the supported
+// flags and failing early if an unknown flag is provided.
+func flagsCmd(pr *external.PluginRequest) external.PluginResponse {
+	pluginResponse := external.PluginResponse{
+		APIVersion: "v1alpha1",
+		Command:    "flags",
+		Universe:   pr.Universe,
+		Flags:      []external.Flag{},
+	}
+
+	// Here is an example of parsing multiple flags from a Kubebuilder external plugin request
+	flagsToParse := pflag.NewFlagSet("flagsFlags", pflag.ContinueOnError)
+	flagsToParse.Bool("init", false, "sets the init flag to true")
+	flagsToParse.Bool("api", false, "sets the api flag to true")
+	flagsToParse.Bool("webhook", false, "sets the webhook flag to true")
+
+	flagsToParse.Parse(pr.Args)
+
+	initFlag, _ := flagsToParse.GetBool("init")
+	apiFlag, _ := flagsToParse.GetBool("api")
+	webhookFlag, _ := flagsToParse.GetBool("webhook")
+
+	// The Phase 2 Plugins implementation will only ever pass a single boolean flag
+	// argument in the JSON request `args` field. The flag will be `--init` if it is
+	// attempting to get the flags for the `init` subcommand, `--api` for `create api`,
+	// `--webhook` for `create webhook`, and `--edit` for `edit`
+	if initFlag {
+		// Add a flag to the JSON response `flags` field that Kubebuilder reads
+		// to ensure it binds to the flags given in the response.
+		pluginResponse.Flags = append(pluginResponse.Flags, external.Flag{
+			Name:    "domain",
+			Type:    "string",
+			Default: "example.domain.com",
+			Usage:   "sets the domain added in the scaffolded initFile.txt",
+		})
+	} else if apiFlag {
+		pluginResponse.Flags = append(pluginResponse.Flags, external.Flag{
+			Name:    "number",
+			Type:    "int",
+			Default: "1",
+			Usage:   "set a number to be added in the scaffolded apiFile.txt",
+		})
+	} else if webhookFlag {
+		pluginResponse.Flags = append(pluginResponse.Flags, external.Flag{
+			Name:    "hooked",
+			Type:    "bool",
+			Default: "false",
+			Usage:   "add the word `hooked` to the end of the scaffolded webhookFile.txt",
+		})
+	} else {
+		pluginResponse.Error = true
+		pluginResponse.ErrorMsgs = []string{
+			"unrecognized flag",
+		}
+	}
+
+	return pluginResponse
+}
